extract shared member row for manage roles and invites
